requirements.txt
Flask
psycopg2-binary
pytest
pytest-timeout
tkinter
